relocate fs package to internal/fs
